requirements.txt
streamlit>=1.34
pillow>=10.0.0
numpy>=1.26.0